# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /swagger:
    x-swagger-pipe: swagger_raw
  /route/list:
    x-swagger-router-controller: routes
    get:
      description: Returns routes list
      operationId: getRouteList
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/Route"
  /route/{id}:
    x-swagger-router-controller: routes
    get:
      description: Returns a Route by id
      operationId: getRoute
      parameters:
        - name: id
          in: path
          description: ID of the route
          required: true
          type: integer
#          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Route"
        "404":
          description: Route not found
          schema:
            $ref: "#/definitions/ErrorResponse"
  /state:
    x-swagger-router-controller: routes
    get:
      description: Returns list of states for specified transport IMEI numbers
      operationId: getState
      parameters:
        - name: imei
          in: query
          description: Transport keys
          type: array
          items:
            type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/State"
        "500":
          description: Server error

# complex objects have schema definitions
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Route:
    properties:
      id:
        type: string
      Number:
        type: string
      Type:
        type: string
      title:
        type: string
      color:
        type: string
      distance:
        type: string
      cost:
        type: string
      last_update:
        type: string
  State:
    properties:
      lat:
        type: string
      lng:
        type: string
      ts:
        type: string
      speed:
        type: string
      azimut:
        type: string
      ignit:
        type: string
      activesim:
        type: string
      sats:
        type: string
      gsmpower:
        type: string
      imei:
        type: string